// shared/src/contracts/manage/media-items/types.d.ts
export interface ManageMediaPerson {
    name: string;
    role?: string;
    thumbUrl?: string;
    profile?: string;
}
export interface ManageMediaExternalId {
    provider: string;
    id: string;
}
export interface ManageMediaItemMetadataRecord {
    title?: string;
    originalTitle?: string;
    sortTitle?: string;
    year?: number;
    overview?: string;
    communityRating?: number;
    genres: string[];
    directors: string[];
    actors: ManageMediaPerson[];
    studios: string[];
    premiered?: string;
    externalIds: ManageMediaExternalId[];
}