// shared/src/contracts/manage/media-items/api/types.d.ts
export interface RawManageMediaPerson {
    name: string;
    role?: string | null;
    thumb_url?: string | null;
    profile?: string | null;
}
export interface RawManageMediaExternalId {
    provider: string;
    id: string;
}
export interface RawManageMediaItemMetadata {
    title?: string | null;
    original_title?: string | null;
    sort_title?: string | null;
    year?: number | null;
    overview?: string | null;
    community_rating?: number | null;
    genres?: string[] | null;
    directors?: string[] | null;
    actors?: RawManageMediaPerson[] | null;
    studios?: string[] | null;
    premiered?: string | null;
    external_ids?: RawManageMediaExternalId[] | null;
}
// ...
export interface RawManageProbeTaskStreamRecord {
    index?: number | null;
    codec_name?: string | null;
    codec_tag?: string | null;
    title?: string | null;
    language?: string | null;
    channels?: number | null;
    channel_layout?: string | null;
    width?: number | null;
    height?: number | null;
    profile?: string | null;
    bit_rate?: number | null;
    bit_depth?: number | null;
    pixel_format?: string | null;
    color_primaries?: string | null;
    color_space?: string | null;
    color_transfer?: string | null;
    aspect_ratio?: string | null;
    average_frame_rate?: number | null;
    real_frame_rate?: number | null;
    dynamic_range_label?: string | null;
    dv_version_major?: number | null;
    dv_version_minor?: number | null;
    dv_profile?: number | null;
    dv_level?: number | null;
    rpu_present_flag?: number | null;
    el_present_flag?: number | null;
    bl_present_flag?: number | null;
    dv_bl_signal_compatibility_id?: number | null;
    hdr10_plus_present_flag?: boolean | null;
    is_default?: boolean | null;
    is_forced?: boolean | null;
}